replace filters with tabs for bridges tvl view
